slack_notifier/requirements.txt
slack-sdk==3.5.1
google-cloud-secret-manager==2.4.0
google-cloud-build==3.2.1
